Upgrade to AGP 9 (Build failed)
This commit refactors the project structure by decoupling the platform-specific application logic from the core UI components, while updating the build system and dependencies.

Key changes include:
- **Project Restructuring**: Split the monolithic `composeApp` into separate `:androidApp`, `:desktopApp`, and a shared `:composeApp` library.
- **Gradle & Plugin Updates**: Upgraded Gradle wrapper to 9.3.1 and AGP to 9.1.0. Introduced `androidMultiplatformLibrary` and `kotlinJvm` plugins.
- **Dependency Upgrades**: Updated major libraries including Kotlin (2.3.20), Compose Plugin (1.10.3), Koin (4.2.0), Ktor (3.4.1), and Room (2.8.4).
- **Build Configuration**: Migrated Android-specific configurations and resources from `composeApp` to the new `androidApp` module. Similarly, moved desktop-specific logic to `desktopApp`.
- **Cleanup**: Removed `BuildKonfig` and `kotlinter` configurations from the shared module build script to simplify the KMP setup.
