+++ androidApp/src/main/AndroidManifest.xml
