+++ androidApp/src/main/res/drawable/ic_launcher_foreground.webp
