+++ androidApp/src/main/res/drawable/logo_splash_screen.webp
