+++ build.gradle.kts
@@ -2,10 +2,13 @@ plugins {
     // this is necessary to avoid the plugins to be loaded multiple times
     // in each subproject's classloader
     alias(libs.plugins.androidApplication) apply false
+    alias(libs.plugins.androidMultiplatformLibrary) apply false
     alias(libs.plugins.androidLibrary) apply false
+    alias(libs.plugins.androidBuiltInKotlin) apply false
     alias(libs.plugins.jetbrainsCompose) apply false
     alias(libs.plugins.compose.compiler) apply false
     alias(libs.plugins.kotlinMultiplatform) apply false
+    alias(libs.plugins.kotlinJvm) apply false
     alias(libs.plugins.kotlinter) apply false
     alias(libs.plugins.buildKonfig) apply false
     alias(libs.plugins.kotlin.serialization) apply false
