+++ gradle/libs.versions.toml
@@ -1,29 +1,43 @@
 [versions]
-agp = "8.13.1"
-androidx-activityCompose = "1.12.0"
-androidx-lifecycle = "2.10.0-alpha05"
+agp = "9.1.0"
+androidx-activityCompose = "1.13.0"
+androidx-lifecycle = "2.10.0"
 buildKonfig = "0.17.1"
-coil = "3.3.0"
+coil = "3.4.0"
 compose-adaptive = "1.3.0-alpha02"
-compose-plugin = "1.10.0-beta02"
+compose-plugin = "1.10.3"
 cryptography = "0.5.0"
-datastore = "1.2.0"
-kotlin = "2.2.21"
+datastore = "1.2.1"
+kotlin = "2.3.20"
 kotlinx-coroutines = "1.10.2"
 kotlin-datetime = "0.7.1"
-kotlinter = "5.3.0"
-koin = "4.1.1"
-ksp = "2.2.20-2.0.4"
-ktor = "3.3.2"
-mockk = "1.14.6"
-navigation-compose = "2.9.1"
-okio = "3.16.4"
+kotlinter = "5.4.2"
+koin = "4.2.0"
+ksp = "2.3.5"
+ktor = "3.4.1"
+mockk = "1.14.9"
+navigation-compose = "2.9.2"
+okio = "3.17.0"
 room = "2.8.4"
 sqlite = "2.6.2"
 splashscreen = "1.2.0"
+coreKtx = "1.18.0"
+junit = "4.13.2"
+junitVersion = "1.3.0"
+espressoCore = "3.7.0"
+appcompat = "1.7.1"
+material = "1.13.0"
+material3 = "1.10.0-alpha05"
 
 [libraries]
 #=== Jetbrains
+compose-runtime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "compose-plugin" }
+compose-foundation = { module = "org.jetbrains.compose.foundation:foundation", version.ref = "compose-plugin" }
+compose-material3 = { module = "org.jetbrains.compose.material3:material3", version.ref = "material3" }
+compose-ui = { module = "org.jetbrains.compose.ui:ui", version.ref = "compose-plugin" }
+compose-components-resources = { module = "org.jetbrains.compose.components:components-resources", version.ref = "compose-plugin" }
+compose-uiToolingPreview = { module = "org.jetbrains.compose.ui:ui-tooling-preview", version.ref = "compose-plugin" }
+compose-uiTooling = { module = "org.jetbrains.compose.ui:ui-tooling", version.ref = "compose-plugin" }
 kotlinx-coroutines = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "kotlinx-coroutines" }
 kotlinx-coroutines-swing = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-swing", version.ref = "kotlinx-coroutines" }
 androidx-lifecycle-viewmodel = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel", version.ref = "androidx-lifecycle" }
@@ -73,13 +87,22 @@ cryptography-provider-apple = { module = "dev.whyoleg.cryptography:cryptography-
 #=== Test
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
 mockk = { module = "io.mockk:mockk", version.ref = "mockk" }
+androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+junit = { group = "junit", name = "junit", version.ref = "junit" }
+androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
+androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
+androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
+androidMultiplatformLibrary = { id = "com.android.kotlin.multiplatform.library", version.ref = "agp" }
 androidLibrary = { id = "com.android.library", version.ref = "agp" }
+androidBuiltInKotlin = { id = "com.android.built-in-kotlin", version.ref = "agp" }
 jetbrainsCompose = { id = "org.jetbrains.compose", version.ref = "compose-plugin" }
 compose-compiler = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
 kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
+kotlinJvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
 kotlinter = { id = "org.jmailen.kotlinter", version.ref = "kotlinter" }
 buildKonfig = { id = "com.codingfeline.buildkonfig", version.ref = "buildKonfig" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
